Move native types to xdr-types.
